requirements.txt
certifi==2020.12.5
chardet==4.0.0
comtypes==1.1.8
idna==2.10
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.19.4
opencv-python==4.5.1.48
Pillow==8.0.1
psutil==5.8.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
Pymem==1.6
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pywin32==300
pywinauto==0.6.8
requests==2.25.1
six==1.15.0
sty==1.0.0b12
urllib3==1.26.2